DONE celebrity search by list all campaigns
EVERYTHING WORKS UP TO HERE !

and don't worry about, if there is no creator supporter, if he hasn't inserted email !
that's because, only if he doesn't insert email initially, then we have no way to know who was initial creator supporter anyways.
as anyone would be able to use same name, and we shouldn't flag it as creator..
only based on his email, if he's left it. If not, then no way to have him there...